requirements.txt
anyio
arabic-reshaper
asgiref
asn1crypto
Brotli
certifi
cffi
charset-normalizer
cryptography
cssselect2
dj-database-url
Django
django-modeltranslation
fonttools
googletrans
gunicorn
h11
h2
hpack
html5lib
httpcore
httpx
hyperframe
idna
lxml
oscrypto
packaging
pillow
pycparser
pydyf
pyHanko
pyhanko-certvalidator
PyMySQL
pypdf
pyphen
python-bidi
python-dateutil
PyYAML
reportlab
requests
six
sniffio
sqlparse
svglib
tinycss2
tinyhtml5
typing_extensions
tzlocal
uritools
urllib3
weasyprint
webencodings
whitenoise
xhtml2pdf
zopfli
Pillow
deep-translator
openpyxl
django-allauth
PyJWT
# جرب إصداراً محدثاً من psycopg2-binary
psycopg2-binary
python-dotenv
# Optional SMS dependencies (uncomment if using SMS services)
# twilio
# boto3
pycairo

# Django REST Framework and API dependencies
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
django-filter
django-ratelimit